requirements.txt
aiosqlite==0.20.0
black==24.8.0
fastapi==0.110.0
huggingface_hub==0.23.1
IPython==8.25.0
isort== 5.13.2
langchain==0.1.11
langchain-community==0.0.25
langchain_openai==0.1.5
langgraph==0.0.51
langsmith==0.1.19
loguru==0.7.2
llama_index==0.10.65
openai==1.16.0
pandas==2.2.2
pgvector==0.2.5
pip-system-certs==4.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pytest==8.2.2
python-dotenv==1.0.1
PyJWT==2.8.0
sqlalchemy==2.0.30
sqlalchemy.orm==1.2.10
tiktoken==0.5.2
uvicorn==0.29.0
